+++ stubdev/cuda_fullsolve_test_11.cc
@@ -22,7 +22,7 @@ GTEST_TEST(KernelTest, FullSolveTest) {
   Sphere* h_spheres = new Sphere[numProblems * numSpheres];
   for (int i = 0; i < numProblems; i++) {
     for (int j = 0; j < numSpheres; j++) {
-      Eigen::Vector3d p;
+      Eigen::Vector3f p;
       // Case 1 - 22 Spheres in the environment
       if (j == 0) {
         p << 0.0, 0.0, 0.0;
@@ -50,10 +50,10 @@ GTEST_TEST(KernelTest, FullSolveTest) {
         // y between -1.0 and -1.5
         // cur ball position
         // int col = i % 20;
-        // p << -2.4 + static_cast<double>(col) * (4.8 / 20.0), -1.7, 0.0;
+        // p << -2.4 + static_cast<float>(col) * (4.8 / 20.0), -1.7, 0.0;
 
-        // double random_angle =
-        //     static_cast<double>(rand()) / RAND_MAX * 2.0 * M_PI;
+        // float random_angle =
+        //     static_cast<float>(rand()) / RAND_MAX * 2.0 * M_PI;
         // p << 0.0 + 0.15 * cos(random_angle), -0.12 + 0.03 *
         // sin(random_angle),
         //     0.0;
@@ -63,21 +63,21 @@ GTEST_TEST(KernelTest, FullSolveTest) {
 
       h_spheres[i * numSpheres + j].center = p;
 
-      h_spheres[i * numSpheres + j].velocity = Eigen::Vector3d::Zero();
+      h_spheres[i * numSpheres + j].velocity = Eigen::Vector3f::Zero();
 
       h_spheres[i * numSpheres + j].mass = 0.17;
 
       if (j == 10) [[unlikely]] {
         // a random aiming point, from (0,0.25) to (0.0,3.5)
-        Eigen::Vector3d random_target(
-            0.0, 0.03 + static_cast<double>(rand()) / RAND_MAX * 0.15, 0.0);
-        Eigen::Vector3d direction = random_target - p;
+        Eigen::Vector3f random_target(
+            0.0, 0.03 + static_cast<float>(rand()) / RAND_MAX * 0.15, 0.0);
+        Eigen::Vector3f direction = random_target - p;
         direction.normalize();
         // scale up the velocity to 8.0 to 20.0,
         // h_spheres[i * numSpheres + j].velocity =
         //     direction * 1.2 +
-        //     static_cast<double>(rand()) / RAND_MAX * 0.5 * direction;
-        h_spheres[i * numSpheres + j].velocity = Eigen::Vector3d(0.0, 1.2, 0.0);
+        //     static_cast<float>(rand()) / RAND_MAX * 0.5 * direction;
+        h_spheres[i * numSpheres + j].velocity = Eigen::Vector3f(0.0, 1.2, 0.0);
 
         h_spheres[i * numSpheres + j].mass = 0.17;
       }
@@ -100,10 +100,10 @@ GTEST_TEST(KernelTest, FullSolveTest) {
       //   // y between -1.0 and -1.5
       //   // cur ball position
       //   // int col = i % 20;
-      //   // p << -2.4 + static_cast<double>(col) * (4.8 / 20.0), -1.7, 0.0;
+      //   // p << -2.4 + static_cast<float>(col) * (4.8 / 20.0), -1.7, 0.0;
 
-      //   double random_angle =
-      //       static_cast<double>(rand()) / RAND_MAX * 2.0 * M_PI;
+      //   float random_angle =
+      //       static_cast<float>(rand()) / RAND_MAX * 2.0 * M_PI;
       //   p << 0.0 + 0.05 * cos(random_angle), -0.12 + 0.03 *
       //   sin(random_angle),
       //       0.0;
@@ -113,19 +113,19 @@ GTEST_TEST(KernelTest, FullSolveTest) {
 
       // h_spheres[i * numSpheres + j].center = p;
 
-      // h_spheres[i * numSpheres + j].velocity = Eigen::Vector3d::Zero();
+      // h_spheres[i * numSpheres + j].velocity = Eigen::Vector3f::Zero();
 
       // h_spheres[i * numSpheres + j].mass = 0.17;
 
       // if (j == 3) [[unlikely]] {
-      //   Eigen::Vector3d random_target(0.0, 0.0, 0.0);
-      //   Eigen::Vector3d direction = random_target - p;
+      //   Eigen::Vector3f random_target(0.0, 0.0, 0.0);
+      //   Eigen::Vector3f direction = random_target - p;
       //   direction.normalize();
       //   // scale up the velocity to 8.0 to 20.0, random
       //   h_spheres[i * numSpheres + j].velocity =
       //       direction * 1.2 +
-      //       static_cast<double>(rand()) / RAND_MAX * 0.5 * direction;
-      //   // h_spheres[i * numSpheres + j].velocity = Eigen::Vector3d(0.0, 5.0,
+      //       static_cast<float>(rand()) / RAND_MAX * 0.5 * direction;
+      //   // h_spheres[i * numSpheres + j].velocity = Eigen::Vector3f(0.0, 5.0,
       //   // 0.0);
 
       //   h_spheres[i * numSpheres + j].mass = 0.17;
@@ -182,7 +182,7 @@ GTEST_TEST(KernelTest, FullSolveTest) {
   auto end = std::chrono::high_resolution_clock::now();
 
   // Calculate the duration
-  std::chrono::duration<double, std::milli> duration = end - start;
+  std::chrono::duration<float, std::milli> duration = end - start;
   std::cout << "Time taken for 800 iterations: " << duration.count()
             << "mili seconds" << std::endl;
